Multiple upload di ulasan

+++ database/migrations/2025_12_02_122127_add_warga_id_to_produk_table.php
@@ -6,27 +6,27 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('produk', function (Blueprint $table) {
-            $table->foreignId('warga_id')
-                ->after('produk_id')
-                ->constrained('warga', 'warga_id')
-                ->cascadeOnDelete();
+            // Tambah kolom hanya jika belum ada
+            if (!Schema::hasColumn('produk', 'warga_id')) {
+
+                $table->foreignId('warga_id')
+                    ->after('produk_id')
+                    ->constrained('warga', 'warga_id')
+                    ->cascadeOnDelete();
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('produk', function (Blueprint $table) {
-            $table->dropForeign(['warga_id']);
-            $table->dropColumn('warga_id');
+            if (Schema::hasColumn('produk', 'warga_id')) {
+                $table->dropForeign(['warga_id']);
+                $table->dropColumn('warga_id');
+            }
         });
     }
 };
